Move InputItem to app-wide shared directory
