+++ app/profile/summary.tsx
@@ -297,9 +297,12 @@ export default function Summary() {
             <Card.Body color="primary.fg" gap={1}>
               <HStack mb="6" gap="3">
                 <Avatar
+                  bgGradient="to-tr"
+                  gradientFrom={{ _dark: "green.600", _light: "green.400" }}
+                  gradientTo={{ _dark: "yellow.600", _light: "yellow.400" }}
+                  gradientVia={{ _dark: "pink.600", _light: "pink.300" }}
                   src={image}
                   name={name}
-                  bg="primary.contrast"
                   variant={"outline"}
                 />
                 <Stack gap="0">
